% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-asep.R
\docType{data}
\name{asep7}
\alias{asep7}
\title{Essay Set 7 from the Hewlett Foundation: Automated Essay Scoring}
\format{
A data frame with 1,569 rows and 28 columns.
}
\source{
\url{https://www.kaggle.com/competitions/asap-aes/}
}
\usage{
asep7
}
\description{
This is the 8th essay set from the Hewlett Foundation Automated Essay Scoring
dataset from Kaggle. The original dataset was provided as part of a competition
to develop an automated scoring algorithm for student-written essays. The
prompt for essay set 7 is:
}
\details{
\emph{Write about patience. Being patient means that you are understanding and
tolerant. A patient person experience difficulties without complaining.}

\emph{Do only one of the following: write a story about a time when you were
patient OR write a story about a time when someone you know was patient
OR write a story in your own way about patience.}

The essay was evaluated by two to three rates on the following six traits:
\enumerate{
\item Ideas
\item Organization
\item Style
\item Conventions
}
}
\seealso{
asep8_rubric
}
\keyword{datasets}
