% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-asep.R
\docType{data}
\name{asep7_rubric}
\alias{asep7_rubric}
\title{Rubric for essay set seventh from the Hewlett Foundation Automated Essay Scoring dataset}
\format{
A data frame with 4 rows and 5 columns. The rows correspond to the four
traits each essay is evaluated on (ideas, organizatino, style, adn conventions)
and the columns correspond to the score zero through three.
}
\source{
\url{https://www.kaggle.com/competitions/asap-aes/}
}
\usage{
asep7_rubric
}
\description{
Rubric for essay set seventh from the Hewlett Foundation Automated Essay Scoring dataset
}
\seealso{
asep8
}
\keyword{datasets}
