% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-asep.R
\docType{data}
\name{asep8}
\alias{asep8}
\title{Essay Set 8 from the Hewlett Foundation: Automated Essay Scoring}
\format{
A data frame with 723 rows and 28 columns.
}
\source{
\url{https://www.kaggle.com/competitions/asap-aes/}
}
\usage{
asep8
}
\description{
This is the 8th essay set from the Hewlett Foundation Automated Essay Scoring
dataset from Kaggle. The original dataset was provided as part of a competition
to develop an automated scoring algorithm for student-written essays. The
prompt for essay set 8 is:
}
\details{
\emph{We all understand the benefits of laughter. For example, someone once said,
"Laughter is the shortest distance between two people." Many other people
believe that laughter is an important part of any relationship. Tell a true
story in which laughter was one element or part.}

The essay was evaluated by two to three rates on the following six traits:
\enumerate{
\item Ideas and Content
\item Organization
\item Voice
\item Word Choice
\item Sentence Fluency
\item Conventions
}
}
\seealso{
asep8_rubric
}
\keyword{datasets}
