% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-asep.R
\docType{data}
\name{asep8_rubric}
\alias{asep8_rubric}
\title{Rubric for essay set eight from the Hewlett Foundation Automated Essay Scoring dataset}
\format{
A data frame with 6 rows and 7 columns. The rows correspond to the six
traits each essay is evaluated on (Ideas and Content; Organization; Voice;
Word Choice; Sentence Fluency; and Conventions) and the columns correspond to
the score six through one.
}
\source{
\url{https://www.kaggle.com/competitions/asap-aes/}
}
\usage{
asep8_rubric
}
\description{
Rubric for essay set eight from the Hewlett Foundation Automated Essay Scoring dataset
}
\seealso{
asep8
}
\keyword{datasets}
