% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Utility function to determine whether the package dependencies are available.}
\usage{
check_dependencies(dependencies)
}
\arguments{
\item{dependencies}{character vector with the package names to check.}
}
\value{
TRUE if all the required packages are available and have been loaded
using the \code{\link[base:ns-load]{base::requireNamespace()}} function.
}
\description{
Utility function to determine whether the package dependencies are available.
}
