% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_module.R
\name{codebook_server}
\alias{codebook_server}
\title{Server for the codebook.}
\usage{
codebook_server(id, qda_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}
}
\description{
Server for the codebook.
}
