% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_module.R
\name{codebook_ui}
\alias{codebook_ui}
\title{UI for the codebook.}
\usage{
codebook_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the codebook.
}
