% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_module.R
\name{coding_server}
\alias{coding_server}
\title{Server text coding.}
\usage{
coding_server(id, qda_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}
}
\description{
Server text coding.
}
