% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_module.R
\name{coding_ui}
\alias{coding_ui}
\title{UI for text coding.}
\usage{
coding_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for text coding.
}
