% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{This palette was created from https://colorbrewer2.org using qualitative
type with 12 classes. There are two schemes given these parameters, the first
12 are from a lighter palette and the second 12 are from a darker palette.}
\format{
An object of class \code{character} of length 24.
}
\usage{
color_palette
}
\description{
This palette was created from https://colorbrewer2.org using qualitative
type with 12 classes. There are two schemes given these parameters, the first
12 are from a lighter palette and the second 12 are from a darker palette.
}
\keyword{datasets}
