% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-daacs.R
\docType{data}
\name{daacs}
\alias{daacs}
\alias{daacs_data}
\alias{daacs_codings}
\alias{daacs_text_responses}
\alias{daacs_rubric}
\title{Sample data from the Diagnostic Assessment and Achievement of College Skills (DAACS)}
\format{
Data frames.

An object of class \code{data.frame} with 311 rows and 8 columns.

An object of class \code{data.frame} with 17 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
}
\source{
\url{https://daacs.net}

\url{https://docs.daacs.net/overview/assessments/writing/}
}
\usage{
daacs_data

daacs_codings

daacs_text_responses

daacs_rubric
}
\description{
Sample data from the Diagnostic Assessment and Achievement of College Skills (DAACS)
}
\section{About DAACS}{


This is a small subset of essays, essay scores, and self-regulated learning
results from the Diagnostic Assessment and Achievement of College Skills (DAACS).
DAACS is a suite of technological and social supports to optimize student
learning. DAACS provides personalized feedback about students' strengths and
weaknesses in terms of key academic and self-regulated learning skills,
linking them to the resources to help them be successful students.

These data are a subset of essays used in ChingYen et al (2023) College
Students' Responses to Automated Feedback on Self-Regulated Learning.
}

\section{\code{daacs_data} Structure}{

\itemize{
\item \code{id} integer
\item \code{qda_text} character
\item \code{content_summary} numeric
\item \code{content_suggestions} numeric
\item \code{organization_structure} numeric
\item \code{organization_transitions} numeric
\item \code{paragraphs_ideas} numeric
\item \code{paragraphs_cohesion} numeric
\item \code{sentences_correct} numeric
\item \code{sentences_complexity} numeric
\item \code{conventions_usage} numeric
\item \code{conventions_punctuation} numeric
\item \code{motivation} numeric
\item \code{strategies} numeric
\item \code{metacognition} numeric
\item \code{selfefficacy} numeric
}
}

\section{\code{daacs_coding} Structure}{

\itemize{
\item \code{coding_id} integer
\item \code{coder} character
\item \code{id} integer
\item \code{text} character
\item \code{codes} character
\item \verb{Content of Essay} character
\item \verb{Does this text represent a} character
\item \verb{Judgment about survey or feedback} character
}
}

\section{\code{daacs_text_responses} Structure}{

\itemize{
\item \code{id} integer
\item \code{stem} character
\item \code{answer} character
\item \code{coder} character
}
}

\seealso{
daacs_codings daacs_text_responses daacs_rubric
}
\keyword{datasets}
