% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_view_module.R
\name{data_view_server}
\alias{data_view_server}
\title{Server for the data view.}
\usage{
data_view_server(id, qda_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see [@ShinyQDA::qda()].}
}
\description{
Server for the data view.
}
