% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_view_module.R
\name{data_view_ui}
\alias{data_view_ui}
\title{UI for the data view.}
\usage{
data_view_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the data view.
}
