% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_module.R
\name{descriptives_ui}
\alias{descriptives_ui}
\title{UI for the descriptive statistics.}
\usage{
descriptives_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the descriptive statistics.
}
