% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_app.R
\name{find_document_column}
\alias{find_document_column}
\title{Find the column with the longest character string.}
\usage{
find_document_column(df)
}
\arguments{
\item{df}{a data.frame to search for the document column.}
}
\value{
the name of the column that is likely to contain the documents (i.e. text).
}
\description{
Find the column with the longest character string.
}
