% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ShinyQDA_app.R
\name{find_primary_key_column}
\alias{find_primary_key_column}
\title{Finds the first column that can be used as the primary key.}
\usage{
find_primary_key_column(df)
}
\arguments{
\item{df}{a data.frame to search for a primary key.}
}
\value{
the name of the first column that could be a primary key or NULL if one could not be found.
}
\description{
This function loops through the given data.frame and returns the name of the
first column it finds that could be used as a primary key. That is, the first
column where the length of unique values equals the number of rows.
}
