% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda_merge.R
\name{get_coding_table}
\alias{get_coding_table}
\title{Returns a table where each row represents a text coded by a coder and each
column represents the number of codes present by that coder.}
\usage{
get_coding_table(qda_data, aggregate_fun = sum)
}
\arguments{
\item{qda_data}{a [ShinyQDA::qda()] object.}

\item{aggregate_fun}{function used to aggregate multiple tags per text per
per coder. See [reshape2::dcast()] for more information.}
}
\description{
Returns a table where each row represents a text coded by a coder and each
column represents the number of codes present by that coder.
}
