% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighter.R
\name{highlighter}
\alias{highlighter}
\title{Highlights the codes within the text using HTML.}
\usage{
highlighter(text, codings, codes, link = TRUE)
}
\arguments{
\item{text}{the text to highlight.}

\item{codings}{a data.frame with the codings.}

\item{codes}{a data.frame with the codes (and colors).}

\item{link}{if TRUE codes will link to a modal dialog to edit them.}
}
\description{
Highlights the codes within the text using HTML.
}
