% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_highlighter.R
\docType{data}
\name{lexicon_colors}
\alias{lexicon_colors}
\title{Color palettes for the various sentiment lexicons}
\format{
An object of class \code{list} of length 4.
}
\usage{
lexicon_colors
}
\description{
Color palettes for the various sentiment lexicons
}
\keyword{datasets}
