% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{missing_dependencies}
\alias{missing_dependencies}
\title{Utility function to determine whether the package dependencies are available.}
\usage{
missing_dependencies(dependencies)
}
\arguments{
\item{dependencies}{character vector with the package names to check.}
}
\value{
character vector of packages not available.
}
\description{
Utility function to determine whether the package dependencies are available.
}
