% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalDialog2.R
\name{modalDialog2}
\alias{modalDialog2}
\title{Adapted from https://stackoverflow.com/questions/63882483/how-to-adjust-shiny-modaldialog-width-to-a-dt-object-to-fully-show-the-table-i}
\usage{
modalDialog2(
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  easyClose = FALSE,
  fade = TRUE
)
}
\description{
Adapted from https://stackoverflow.com/questions/63882483/how-to-adjust-shiny-modaldialog-width-to-a-dt-object-to-fully-show-the-table-i
}
