% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_app.R
\name{new_app}
\alias{new_app}
\title{Create a new ShinyQDA application}
\usage{
new_app(
  name = "My_ShinyQDA",
  dir = ifelse(interactive(), rstudioapi::selectDirectory(caption =
    "Select location for ShinyQDA application"), getwd()),
  qda_data,
  id_column = find_primary_key_column(qda_data),
  text_column = find_document_column(qda_data),
  initialize_sentiment_dictionaries = TRUE,
  run_app = interactive()
)
}
\arguments{
\item{name}{name of the application directory.}

\item{dir}{directory where the application will be created.}

\item{qda_data}{a data.frame with the text data.}

\item{id_column}{the name of the column in \code{qda_data} for the primary key.}

\item{text_column}{the name of the column in \code{qda_data} containing the text data.}

\item{initialize_sentiment_dictionaries}{whether to copy the sentiment
dictionaries into the app directory.}

\item{run_app}{if TRUE the Shiny application will be started once it has been created.}
}
\value{
Returns a \code{\link[=qda]{qda()}} object.
}
\description{
Create a new ShinyQDA application
}
