% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ShinyQDA_app.R
\name{new_ShinyQDA_app}
\alias{new_ShinyQDA_app}
\title{Creates a new ShinyQDA application}
\usage{
new_ShinyQDA_app(
  name = "My_ShinyQDA",
  dir = ifelse(interactive(), rstudioapi::selectDirectory(caption =
    "Select location for ShinyQDA application"), getwd()),
  qda_data,
  id_column = find_primary_key_column(qda_data),
  document_column = find_document_column(qda_data),
  initialize_sentiment_dictionaries = TRUE
)
}
\description{
Creates a new ShinyQDA application
}
