% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda_datatable.R
\name{qda_datatable}
\alias{qda_datatable}
\title{Wrapper to \code{datatable} function.}
\usage{
qda_datatable(df)
}
\arguments{
\item{df}{a data.frame.}
}
\description{
This is a wrapper to the \code{\link[DT:datatable]{DT::datatable()}} function. The primary purpose is
to centralize the options to \code{datatable} so that all tables in the ShinyQDA
app are consistent.
}
