% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda_view_module.R
\name{qda_view_ui}
\alias{qda_view_ui}
\title{UI for the raw data view.}
\usage{
qda_view_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
This will display at tab group for each table in the \code{\link[=qda]{qda()}} data object.
}
