% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda_merge.R
\name{question_responses_to_data_frame}
\alias{question_responses_to_data_frame}
\title{Utility function to convert code and text questions to a wide data.frame.}
\usage{
question_responses_to_data_frame(questions, responses, id_col)
}
\description{
Utility function to convert code and text questions to a wide data.frame.
}
