% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions_module.R
\name{questions_server}
\alias{questions_server}
\title{Server the raw data view.}
\usage{
questions_server(id, qda_data, page_length = 20)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}
}
\description{
Server the raw data view.
}
