% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions_module.R
\name{questions_ui}
\alias{questions_ui}
\title{UI for the raw data view.}
\usage{
questions_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
This will display at tab group for each table in the [qda()] data object.
}
