% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_module.R
\name{reliability_ui}
\alias{reliability_ui}
\title{UI for the inter-rater reliability analysis.}
\usage{
reliability_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the inter-rater reliability analysis.
}
