% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubric_module.R
\name{rubric_edit_server}
\alias{rubric_edit_server}
\title{Server logic for the rubric.}
\usage{
rubric_edit_server(id, qda_data, page_length = 20)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}

\item{page_length}{the default number of rows to show in the table.}
}
\description{
Server logic for the rubric.
}
