% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubric_module.R
\name{rubric_edit_ui}
\alias{rubric_edit_ui}
\title{UI for the rubric.}
\usage{
rubric_edit_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the rubric.
}
