% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_module.R
\name{sentiment_server}
\alias{sentiment_server}
\title{Server for the sentiment analysis.}
\usage{
sentiment_server(id, qda_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}
}
\description{
Server for the sentiment analysis.
}
