% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_module.R
\name{sentiment_ui}
\alias{sentiment_ui}
\title{UI for the sentiment analysis}
\usage{
sentiment_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the sentiment analysis
}
