% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{shiny_server}
\alias{shiny_server}
\title{Shiny Server for QDA}
\usage{
shiny_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
Shiny Server for QDA
}
