% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{shiny_ui}
\alias{shiny_ui}
\title{Shiny User Interface for QDA}
\usage{
shiny_ui(request, debug = TRUE)
}
\arguments{
\item{request}{a Shiny request object.}

\item{debug}{parater passed to \code{\link[shinyjs:useShinyjs]{shinyjs::useShinyjs()}}}
}
\description{
Shiny User Interface for QDA
}
