% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyQDA-package.R
\docType{package}
\name{ShinyQDA-package}
\alias{ShinyQDA}
\alias{ShinyQDA-package}
\title{ShinyQDA: R Package and Shiny Application for the Analysis of Qualitative Data}
\description{
This package provides tools to conduct qualitative data analysis (QDA) in R. The Shiny application provide an interactive way to code qualitative data. Utility functions are provided to then analyze the data in R. Data is stored in a SQL database and works with any database that works with the DBI R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbryer/ShinyQDA}
  \item \url{https://jbryer.github.io/ShinyQDA/}
  \item Report bugs at \url{https://github.com/jbryer/ShinyQDA/issues}
}

}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{internal}
