% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run.R
\name{shinyQDA}
\alias{shinyQDA}
\title{Run the ShinyQDA application locally.}
\usage{
shinyQDA(qda_data, ...)
}
\arguments{
\item{...}{other parmaeters passed to [shiny::runApp].}

\item{df}{data.frame containing the essays}
}
\description{
Run the ShinyQDA application locally.
}
