% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_truncate}
\alias{text_truncate}
\title{Returns the first words that fit within the given number of characeters.}
\usage{
text_truncate(value, width = 50)
}
\arguments{
\item{value}{character to truncate}

\item{width}{maximum number of characters the resulting string should be.}
}
\description{
Returns the first words that fit within the given number of characeters.
}
