% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{text_wrap_mod}
\alias{text_wrap_mod}
\title{Wrap label text.}
\usage{
text_wrap_mod(value, width = 25)
}
\arguments{
\item{value}{vector (converted using \code{\link{as.character}}) to be wrapped.}

\item{width}{the maximum width of each line in characters.

Adapted from https://github.com/hadley/ggplot2/wiki/labeller}
}
\description{
Wrap label text.
}
