% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_module.R
\name{topic_modeling_server}
\alias{topic_modeling_server}
\title{Server for the sentiment analysis.}
\usage{
topic_modeling_server(id, qda_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{qda_data}{QDA data object, see \code{\link[=qda]{qda()}}.}
}
\description{
Much of the code and analysis was adapted from https://ladal.edu.au/topicmodels.html
}
