% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_module.R
\name{topic_modeling_ui}
\alias{topic_modeling_ui}
\title{UI for the topic modeling analysis}
\usage{
topic_modeling_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\description{
UI for the topic modeling analysis
}
