% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{highlighter}
\alias{highlighter}
\title{Converts the text to something more reasonable to use as an input ID.}
\usage{
highlighter(text, codings, codes)
}
\arguments{
\item{text}{the text to highlight.}

\item{codings}{a data.frame with the codings.}

\item{codes}{a data.frame with the codes (and colors).}
}
\description{
Converts the text to something more reasonable to use as an input ID.
}
