% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_highlighter.R
\docType{data}
\name{lexicon_colors}
\alias{lexicon_colors}
\title{Color palettes for the various sentiment lexicons}
\format{
a list containing character vectors.
}
\usage{
lexicon_colors
}
\description{
This defines the colors used for various sentiment analyses. Each element of
the list corresponds to a sentiment dictionary with each element a character
vector where the names correspond to the sentiment words and the values
are hex colors.
}
\keyword{datasets}
