% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ShinyQDA_app.R
\name{new_ShinyQDA_app}
\alias{new_ShinyQDA_app}
\title{Creates a new ShinyQDA application}
\usage{
new_ShinyQDA_app(
  name = "My_ShinyQDA",
  dir = ifelse(interactive(), rstudioapi::selectDirectory(caption =
    "Select location for ShinyQDA application"), getwd()),
  qda_data,
  id_column = find_primary_key_column(qda_data),
  document_column = find_document_column(qda_data),
  initialize_sentiment_dictionaries = TRUE
)
}
\arguments{
\item{name}{name of the application directory.}

\item{dir}{directory where the application will be created.}

\item{qda_data}{a data.frame with the text data.}

\item{id_column}{the name of the column in \code{qda_data} for the primary key.}

\item{document_column}{the name of the column in \code{qda_data} containing the text data.}

\item{initialize_sentiment_dictionaries}{whether to copy the sentiment
dictionaries into the app directory.}
}
\description{
Creates a new ShinyQDA application
}
