% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda.R
\name{qda}
\alias{qda}
\title{Data object for qualitative data analysis}
\usage{
qda(
  file,
  df,
  id_col = names(df)[1],
  text_col = names(df)[2],
  codes = character(),
  auto_save = TRUE
)
}
\description{
Data object for qualitative data analysis
}
