% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qda.R
\name{qda}
\alias{qda}
\title{Data object for qualitative data analysis.}
\usage{
qda(file, users_passphrase = "ShinyQDA")
}
\description{
This function creates an object to manipulate data for qualitative analysis.
The data is stored in a SQLite database by default, but can be configures to
work with any SQL database through the `DBI` R package interface. Although
you can manipulate the SQL database directly, we recommend using the methods
(functions) defined in the returned object. This will ensure data integrity.
}
\details{
* `get_text(id)`
* `add_coding(id, text = NA, start = NA, end = NA, codes = NA, coder = NA)`
}
\section{Methods}{

\describe{
   \item{\code{add_text(df, id_col, text_col, overwrite = TRUE, append = FALSE)}}{Adds text data to the object. There needs to be at least two columns, the column containing the text and a column containing a unique identifier (i.e. primary key). You can include any other columns that will be available in the Shiny app for analysis.}
}
}

