% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_highlighter.R
\name{sentiment_highlighter}
\alias{sentiment_highlighter}
\title{Returns an HTML character highlighting the sentiments.}
\usage{
sentiment_highlighter(text, lexicon = c("nrc", "bing", "loughran", "afinn"))
}
\arguments{
\item{thetext}{the text to highlight.}
}
\description{
Returns an HTML character highlighting the sentiments.
}
