% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{shiny_ui}
\alias{shiny_ui}
\title{Shiny User Interface for QDA}
\usage{
shiny_ui()
}
\arguments{
\item{title}{title displayed in the top left of the Shiny app.}
}
\description{
Shiny User Interface for QDA
}
