% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run.R
\name{shinyQDA}
\alias{shinyQDA}
\title{Run the ShinyQDA application locally.}
\usage{
shinyQDA(
  qda_data_file,
  authenticate = !interactive(),
  enable_admin = TRUE,
  keep_token = TRUE,
  ...
)
}
\arguments{
\item{qda_data_file}{file name to save results. See [ShinyQDA::qda()] for more information.}

\item{authenticate}{use [shinymanager::secure_app()].}

\item{enable_admin}{enable admin mode for `shinymanager`.}

\item{keep_token}{Logical, keep the token used to authenticate in the URL,
it allow to refresh the application in the browser, but careful the
token can be shared between users.}

\item{...}{other parameters passed to [shiny::runApp()] and [shiny::shinyApp()].}
}
\description{
Run the ShinyQDA application locally.
}
