% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run.R
\name{shinyQDA}
\alias{shinyQDA}
\title{Run the ShinyQDA application locally.}
\usage{
shinyQDA(
  qda_data_file = "ShinyQDA.sqlite",
  authenticate = !interactive(),
  enable_admin = TRUE,
  keep_token = TRUE,
  users_passphrase = "ShinyQDA",
  ...
)
}
\arguments{
\item{qda_data_file}{file name to save results. See \code{\link[=qda]{qda()}} for more information.}

\item{authenticate}{use \code{\link[shinymanager:secure-app]{shinymanager::secure_app()}}.}

\item{enable_admin}{enable admin mode for \code{shinymanager}.}

\item{keep_token}{Logical, keep the token used to authenticate in the URL,
it allow to refresh the application in the browser, but careful the
token can be shared between users.}

\item{users_passphrase}{passphrase to pass to \code{shinymanager}.}

\item{...}{other parameters passed to \code{\link[shiny:runApp]{shiny::runApp()}} and \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Run the ShinyQDA application locally.
}
